+++ web-component/lib/render.ts
@@ -4,7 +4,7 @@ export function renderHTML(): string {
     <div class="popover-backdrop" part="backdrop" aria-hidden="true"></div>
     <div class="popover-content" part="content" tabindex="-1">
         <div class="drag-handle" part="drag-handle" aria-hidden="true"><span></span></div>
-        <slot></slot>
+        <slot part="content-slot"></slot>
     </div>
   </div>
   `;